Implement logic and Fix bugs in  user, auth, company, department, ehist and absences controllers

+++ backend/src/controllers/campanyContoller.js
@@ -1,37 +1,24 @@
 const { Company } = require('../../models/index');
+const DataAccesManagement = require('../utils/dataAccessManagement');
+const findUserByToken = require('../utils/findUserByToken');
 
 class CompanyController {
   static async getCompany(req, res) {
     // #swagger.tags = ['Company']
-    const companies = await Company.findAll();
-    return res.status(200).send({ message: 'Companies List', companies });
-  }
 
-  static async createCompany(req, res) {
-     // #swagger.tags = ['Company']
     try {
-      const { name, address } = req.body;
-      const oldCompany = await Company.findOne({
-        where: {
-          name,
-        },
-      });
-      if (oldCompany) {
-        return res.status(400).send({ message: 'Company already exists' });
-      }
-      const company = await Company.create({
-        name,
-        address,
-      });
-      return res
-        .status(201)
-        .json({ message: 'Company created successfully', company });
-    } catch (error) {
-      console.error('Error creating company:', error);
-      return res
-        .status(500)
-        .json({ message: 'Error creating company', error: error.message });
+      const token = req.headers.authorization;
+
+      const loggedInUser = await findUserByToken(token);
+  
+      const companies = await Company.findAll();
+      const data = DataAccesManagement.companyAccessManager(loggedInUser, companies);
+
+      return res.status(200).send({ message: 'Companies fetched successfully!', companies: data });
+    }catch(err) {
+      return res.status(500).send({ message: 'Failed to fetch companies', error: err });
     }
+   
   }
 
   static async updateCompany(req, res) {
@@ -39,6 +26,24 @@ class CompanyController {
     try {
       const { id } = req.params;
       const { name, address } = req.body;
+
+      const token = req.headers.authorization;
+
+      const loggedInUser = await findUserByToken(token);
+
+      const company = await Company.findOne({
+        where: {
+          id,
+        },
+      });
+
+      const data = DataAccesManagement.companyAccessManager(loggedInUser, [company.dataValues])
+
+      if(data.length === 0) {
+        return res.status(400).send({
+          message: 'You are not authorized to perform this action!',
+        });
+      }
 
       const updatedCompany = await Company.update(
         { name, address },
@@ -57,6 +62,24 @@ class CompanyController {
      // #swagger.tags = ['Company']
     try {
       const { id } = req.params;
+      const token = req.headers.authorization;
+
+      const loggedInUser = await findUserByToken(token);
+
+      const company = await Company.findOne({
+        where: {
+          id,
+        },
+      });
+
+      const data = DataAccesManagement.companyAccessManager(loggedInUser, [company.dataValues])
+
+      if(data.length === 0) {
+        return res.status(400).send({
+          message: 'You are not authorized to perform this action!',
+        });
+      }
+
       await Company.destroy({ where: { id } });
       return res
         .status(200)
@@ -69,8 +92,15 @@ class CompanyController {
   static async getAllCompanies(req, res) {
      // #swagger.tags = ['Company']
     try {
+      const token = req.headers.authorization;
+
+      const loggedInUser = await findUserByToken(token);
+
       const companies = await Company.findAll();
-      return res.status(200).send({ companies });
+
+     const data = DataAccesManagement.companyAccessManager(loggedInUser, companies);
+
+      return res.status(200).send({ message: 'Companies fetched successfully', companies: data });
     } catch (error) {
       return res
         .status(500)
@@ -82,8 +112,26 @@ class CompanyController {
      // #swagger.tags = ['Company']
     try {
       const { id } = req.params;
-      const company = await Company.findByPk(id);
-      return res.status(200).send({ company });
+
+      const token = req.headers.authorization;
+
+      const loggedInUser = await findUserByToken(token);
+
+      const company = await Company.findOne({
+        where: {
+          id,
+        },
+      });
+
+      const data = DataAccesManagement.companyAccessManager(loggedInUser, [company.dataValues])
+
+      if(data.length === 0) {
+        return res.status(400).send({
+          message: 'You are not authorized to perform this action!',
+        });
+      }
+
+      return res.status(200).send({ message: 'Company fetched successfully', company });
     } catch (error) {
       return res.status(500).send({ message: 'Error fetching company', error });
     }
@@ -93,12 +141,12 @@ class CompanyController {
   static async updateCompanyStatus(req, res) {
      // #swagger.tags = ['Company']
     try {
-      const { name } = req.query;
+      const { id } = req.query;
       const { status } = req.body;
 
       const company = await Company.findOne({
         where: {
-          name,
+          id,
         },
       });
 
@@ -121,10 +169,9 @@ class CompanyController {
           company: updatedCompany,
         });
     } catch (error) {
-      console.error(error);
       return res
         .status(500)
-        .send({ message: 'Error approving company', error });
+        .send({ message: 'Error updating company', error });
     }
   }
 }
